Added a new void QXmlStreamWriter::writerCharacters(const QString& text, const int identOffset = 0) member to QXmlStreamWriter class.
When using QXmlStreamWriter::writerCharacters() to write out multi-line plain-text elements to the XML Stream the indent is reset to zero on writing a "\n". This upsets the human readability of the resulting XML.

In order to mitigate this I have exposed a new function which preserves indentation and therefore readability as:

void QXmlStreamWriter::writeIndentedCharacters(const QString& text, const int indentOffset = 0)

This behaves in a similar way to writeCharacters() except that it works out what the current indent level is and prepends its QString equivalent offset by the relative indentOffset parameter. Below is a typical use case and resulting XML:

    stream.writeStartElement("DataLabels");
        stream.writeAttribute("count", "4");
        stream.writeAttribute("independantVars", "2");
        stream.writeCharacters("\n");
        for (int i = 0; i < 2; ++i)
            stream.writeIndentedCharacters(surface.xLabels().at( i ) + "\n");
        stream.writeIndentedCharacters(surface.valLabel() + "\n" + surface.dvalLabel() + "\n");
        stream.writeIndentedCharacters("" ,-1);
    stream.writeEndElement(); // DataLabels

...this results in the following XML snippet, assuming autoFormatting == true and setAutoFormattingIndent is as its default value...

    <DataLabels count="4" independantVars="2">
        stress_x
        stress_y
        pmd
        error_pmd
    </DataLabels>

Reviewed-by: pending