Make getters for staticContentsSupport and partialUpdateSupport virtual
When QRuntimeGraphicsSystem is asked for its support for static contents
of partial updates it should return the value for the currently running
wrapped graphicssystem. As the getters have not been virtual so far this
could not be implemented.
Additionally the setters have been removed as these values are not
supposed to be set from the outside. Only the graphicssystems itself
knows what it supports. If the default values should be changed the
methods should be overwritten.