Copy/Cut/Paste functionality to FEP
Added copy/cut/paste functionality to FEP.FEP copy/cut/paste feature is implemented
as a s60 plugin. The feature is now supported for Symbian^3, 5.0 and 3.2.
Supported QWidgets are QLineEdit,QTextEdit and QPlainTextEdit. From the options
menu, user can copy or cut text (if text is selected) or paste text (if there is
text on the clipboard). Another way of copy/pasting is from the softkeys.
With a long press of the left shift, copy and/or paste can be selected from the
softkeys. User can also select text with long press of the left shift and arrow
keys.

Reviewed-by: Janne Koskinen