Prevent version numbers from being squashed.
If VERSION=1.2.3.4 is specified in the pro file, qmake creates a
"/VERSION:1.234" compiler option. I.e. everything after the first dot
is treated as a single number. This can create a problem since the
compiler will fail if one of the dotted number is greater than 65535.
The compiler will also perform this transformation automatically when
compiling and truncate the number to be small than 65535.