QTBUG-8215 : Fixed deployment problems with MSVC 2005&2008 SP1 See http://bugreports.qt.nokia.com/browse/QTBUG-8215 and http://bit.ly/eLK5Fk
Notably, DLLs bundled with MSVC Expressi SP1  releases are only the one
from the SP1, so it is a good idea to force binding to the latest
release of the runtime to avoid deployment problems.

Keep in mind that it is easier to bundle a private runtime for the
software in an setup EXE rather than lauching a runtime setup from the
software setup.

Reviewed-by: pending