fix for qatomic_{i386,sparc,x86_64}.h for Solaris 10 with Sun Studio 12 C++
This problem was worked out with the help of Thiago Macieria on fixing
an issue on Solaris 10 with Sun Studio 12 C++ compiler, that appeared
between qt-4.7.1 and qt-4.7.2. It was misdiagnosed as a cast issue in
a call from qdbusintegrator.cpp.  This really is an issue in the threads
layer with the wrong types being defined for the fields in the call
to q_atomic_test_and_set_ptr()