Fix calling a PL/SQL stored procedure with an empty array
As per the documentation, the way to do this is to add a null QVariant
of the correct type, to the list. The code in qsql_oci uses that to
detect the column type, and sets the "null" indicator correctly, but
did not change the "curelep" value (actual number of elements) to 0,
so the stored procedure received an array with 1 null item as input.