Remove requirement that Xlib.h not be included before qt_x11_p.h
The qt_x11_p.h header was requiring that the Xlib.h header not be
included before it, but there are situations that arise when using the
LSB compiler where this is unavoidable. The logic within qt_x11_p.h
which leads to this requirement is related to defining the function
prototypes for XRegisterIMInstantiateCallback,
XUnregisterIMInstantiateCallback and XSetIMValues, but the logic is not
necessary when the LSB compilers are used because the function
prototypes provided by the LSB headers are fully known for that
compiler. Thus, the restriction can be removed and the headers allowed
to define these function prototypes in the normal way. The patch to
qt_x11_p.h accomplishes this while preserving the previous behaviour for
all compilers other than the LSB compiler.

The patch to qximinputcontext_x11.cpp fixes a related misuse of these
functions which were relying on incorrect prototypes that were
previously being defined in qt_x11_p.h. These functions have a long
history of some of their parameters being given incorrect types, with
the confusion usually being between XIM and XID, or XIMProc and XIDProc.
The Xlib specification is actually incorrect and this has been generally
accepted by the wider development community. Mainstream linux
distributions have had the fix in place for years, but the Xlib
specification still hasn't been updated. This is confirmed by
discussions on the LSB mailing list also. The patch to
qximinputcontext_x11.cpp makes the code there use the corrected types,
which for the LSB compiler will make the code also match the correct
headers that are part of the LSB SDK. The patch will change no behaviour
from current when using any compiler other than the LSB compiler.

Task-number: QTBUG-16404