Correct possible bug in equal image contents for index8.
When an image is written to a file, it loose the rgb decomposition.
In the test the index8 image returns for pixel(x,y) 0xFFFF0000 because
it's an RGBA image filled with red.
When this pixel is written to disk, its value is 0x57.
When the file is re-read, the only thing we can do is to regenerate a
color table where the 0x57th index is the gray color 0x57.
So the re-read pixel(x,y) is 0xff575757.