Fix don't try to paint on an empty pixmap.
This fixes the problem that when drawItemHelper is called with
widget==NULL then we would end with an empty/invalid QRect
viewRect on line 4553 which is not handled by the intersects-check
on line 4554 cause widget==NULL what can then result in line 4606
on the deviceRect becoming invalid too (cause it does not intersect
with the empty viewRect) what then can result on creating an empty
(aka isNull) QPixmap on line 4621 (4623 with the patch) and then we
would try to call _q_paintIntoCache with that empty QPixmap. This
results then in line 4336 that the pixmapPainter.begin fails.

The patch fixes that by making sure we don't try to paint to an
empty pixmap. This also fixes KDCH-642 (bug related to KDChart where
we earned tons of QPainter begin/end-mismatches).