Ignore custom tooltip palette in QGtkStyle.
Previously if a custom tooltip palette was set with QToolTip::setPalette(),
QGtkStyle would ignore the background color of the custom palette but use
the foreground color. This could lead to unreadable tooltips.

This happens for example with all KDE applications running on GNOME, as kdelibs
setup a custom tooltip palette.

Taking both background and foreground colors from the GTK theme ensures
tooltips are always readable.