Fix bug QTBUG-10330: Problem with "ddd" format in QDateTime::fromString()
Bug occurs when both day of week and day are specified because
the code does not correctly resolve conflicts. Without specifying
date and month, code assumes Jan, 1900. However, day may not match
day of week for default year and month. Commit changes the conflict
resolution such that the first available year or month (depending
on which is free) is found that resolves the conflict. Commit also
appends existing autotest with new dates exhibiting the correct
behavior in previously conflicting cases.