Fix using custom paper sizes on Mac
This solves the bug with using custom paper sizes on Mac, if a QSize
was specified then with a Unit it would not respect this on Mac as it
would not convert it correctly.  Also paperSize() would return the
wrong value.

There is already a test for this in tst_QPrinter::testCustomPageSizes()

Task-number: QTBUG-18723