QTexLayout - Avoid clipping the text line too much when we have selected text
This solves a particular clipping problem.  Sometimes text can be
taller than the height of a line.  This can happen in unicode when using
composite characters.  If we only draw the selected text once, we end up
clipping off the text when it is higher or lower than the line.