Fix a problem with mouse hover after drag and drop
In Win7 if you drag and drop from Qt app to outside the app. Then
move the mouse back over the Qt app and hover over and off a button.
The button highlight will stay on even you are not over the button.

By debugging you can see that the button (or any other widget) will
never receive a leave event and that when moving inside a widget you
will constanly get enter events from the first widget you come in
contact with when moving the mouse back over the app.

Task-number: QTBUG-16823