Early construction of status pane and softkeys for Symbian boosted apps
QtAppBooster can partially start apps and hide them in the background
before they are requested by the user. This gives an apparently faster
start for these apps.

Qt is now detecting these app starts and bringing the construction of
the status pane and softkeys forward so that they happen in the
background. This cuts their construction time off the time between the
user clicking on the app and it being ready to use. On a Nokia E7, this
gain was measured at an average of 127ms.

Task-number: QT-4933
Reviewed-by: Sami Merila