Fix memory leak in QDomDocument DTD notation declaration handler
The created notation node's reference count needs to be decremented to 0 before it is
added as a child, because appendChild will increment the reference count to correct
value of 1. Also added autotest DTDNotationDecl to tst_qdom to expose the leak when
executed under valgrind memcheck. There was no previous test coverage for the notation
declarations in DTD.

PMO 285904

Signed-off-by: Sami Rosendahl <sami.rosendahl@reaktor.fi>