Deactivating an inactive item should do (almost) nothing.
The bug fixed by this task causes the last panel, which can be
any panel depending on the situation, to gain input focus when
an already-inactive item is deactivated by calling setActive(false).
This is obviously wrong.

It's also wrong that setActive(isActive()) is a noop when the
item is already inactive. It's essential that the explicit- and
wantsActive flags are set correctly, so that calling setActive(false)
on an item that's part of an inactive scene (f.ex.) hinders that
item in gaining automatic activation when the scene is activated.

Task-number: QTBUG-12966
Reviewed-by: James Perrett <jamperre@cisco.com>