Forbid open input panel when activeFocusOnPress is set.
Details: You can cheat activeFocusOnPress when you have two input
fields, one with activeFocusOnPress which has been set to true, one with
activeFocusOnPress which has been set to false. Click on the first input
field to open the panel, then click on the second field and observe that the
panel keeps open. This can cause problems in the context of Qt-Components.