Don't use QSettings to "cache" plugin information
This is the 4.8 equivalent of the qtbase commit by Bradley Hughes.
See below.

In addition to the notes below, this change is part of a larger
change to eliminate usage of QSettings where it is unnecessary in
order to make it easier to get Qt applications accepted into the
Mac app store.

From qtbase/8ed47d961dc7e6f161030654d11cd330a542eadf:

The main rationale of the cache was to examine the plugin's build-key
before loading it. Now that the build-key has been removed, the cache
has lost its usefulness.

This is part of a larger push to not use QSettings for Qt specific
settings or caches.

See also:
http://lists.qt.nokia.com/pipermail/qt5-feedback/2011-August/000892.html
http://lists.qt.nokia.com/pipermail/qt5-feedback/2011-August/000960.html
http://lists.qt.nokia.com/pipermail/qt5-feedback/2011-August/000907.html
http://lists.qt.nokia.com/pipermail/qt5-feedback/2011-August/000904.html