Fix clipping of QTextList decorators
List decorators may be clipped if you set a large font size and/or
small indent for a QTextList. The fix is to prevent clipping by
moving list decorators and items to left (or to right in case of
right to left layouts) so that the list decorator is always painted
inside the layout.

Task-number: QTBUG-5111