QTextEngine - Normalize all text to KC composition
Call QString::normalize(QString::NormalizationForm_KC) on all strings
that are passed to harfbuzz to be rendered.  This is because this
version of harfbuzz cannot normalize the text but requires it to be NFC
for it to be rendered correctly.

Note that we use QString::NormalizationForm_KC, Compatibility
Decomposition followed by Canonical Composition. This is because
experimentation showed that strings like ซ้ำ would only be rendered
correct with KC but not C.