Fix crash in QDBusDemarshaller QStringList extraction
QDbusArgument QStringList extraction operator and QDBusDemarshaller that
implements the extraction do not check the type of the extracted value.
When extracting a QStringList and the value actually is e.g. an array of
bytes the string list extraction will crash as it interprets the bytes as
char pointers.

The fix adds DBus type checks to QDBusArgument QStringList extraction
operator implementations.
The checks are as permissive as possible provided crashes are avoided.

Previously supported functionality of extracting a DBus struct with only
string fields in it as a QStringList is maintained. Support will be
removed in Qt5.

PMO 291111
Task-number: QTBUG-22840